+++ Modelador/src/primitivas/Punto.java
@@ -6,41 +6,41 @@ package primitivas;
  *
  */
 public class Punto {
-	private int x;
-	private int y;
-	private int z;
+	private double x;
+	private double y;
+	private double z;
 	
-	public Punto(int x, int y, int z) {
+	public Punto(double x, double y, double z) {
 		this.x = x;
 		this.y = y;
 		this.z = z;
 	}
 	
-	public Punto(int x, int y) {
+	public Punto(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 
 	
-	public int getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(double y) {
 		this.y = y;
 	}
-	public int getZ() {
+	public double getZ() {
 		return z;
 	}
-	public void setZ(int z) {
+	public void setZ(double z) {
 		this.z = z;
 	}
 	public double distancia(Punto b) {
