+++ Modelador/src/primitivas/Recorte.java
@@ -12,34 +12,34 @@ public class Recorte {
 	}
 	
 	
-	public int []recorteHoriz(int []linea, int max_min) {
+	public double []recorteHoriz(double []linea, double max_min) {
 		if(Math.abs(linea[0]-linea[2])<2)
-			return new int[]{-1,-1};
-		int x = (linea[0]+linea[2])/2;
-		int y = (linea[1]+linea[3])/2;
+			return new double[]{-1,-1};
+		double x = (linea[0]+linea[2])/2;
+		double y = (linea[1]+linea[3])/2;
 		
 		if(x<max_min)
-			return recorteHoriz(new int[]{x,y, linea[2], linea[3]}, max_min);
+			return recorteHoriz(new double[]{x,y, linea[2], linea[3]}, max_min);
 		if(x>max_min)
-			return recorteHoriz(new int[]{linea[0],linea[1], x, y}, max_min);
+			return recorteHoriz(new double[]{linea[0],linea[1], x, y}, max_min);
 		else
-			return new int[]{x,y};
+			return new double[]{x,y};
 	}
-	public int []recorteVert(int []linea, int max_min) {
+	public double []recorteVert(double []linea, double max_min) {
 		if(Math.abs(linea[1]-linea[3])<2)
-			return new int[]{-1,-1};
-		int x = (linea[0]+linea[2])/2;
-		int y = (linea[1]+linea[3])/2;
+			return new double[]{-1,-1};
+		double x = (linea[0]+linea[2])/2;
+		double y = (linea[1]+linea[3])/2;
 		
 		if(y<max_min)
-			return recorteVert(new int[]{x,y, linea[2], linea[3]}, max_min);
+			return recorteVert(new double[]{x,y, linea[2], linea[3]}, max_min);
 		if(y>max_min)
-			return recorteVert(new int[]{linea[0],linea[1], x, y}, max_min);
+			return recorteVert(new double[]{linea[0],linea[1], x, y}, max_min);
 		else
-			return new int[]{x,y};
+			return new double[]{x,y};
 	}
 	
-	public  int puntosDentro(int []linea, int []ventana) {
+	public  int puntosDentro(double []linea, double []ventana) {
 		int cont=0;
 		if((linea[0]>=ventana[0])&&(linea[0]<=ventana[2])) {
 			if((linea[1]>=ventana[1])&&(linea[1]<=ventana[3])) {
@@ -53,7 +53,7 @@ public class Recorte {
 		}
 		return cont;
 	}
-	public  boolean  puntosDentroRango(int []linea, int []ventana) {
+	public  boolean  puntosDentroRango(double []linea, double []ventana) {
 		if((linea[0]>=ventana[0])&&(linea[0]<=ventana[2])) {			
 			return true;
 		}
